+++ src/sls/test.ts
@@ -1,5 +1,4 @@
-import type { UNIX_TIMESTAMP } from '../types/common.ts';
-import type { ALIYUN_SLS_ENDPOINT } from './endpoint.ts';
+import { getLogs } from './get-logs.ts';
 import { queryLog } from './mod.ts';
 
 Deno.test('queryLog', async () => {
@@ -20,85 +19,8 @@ Deno.test('queryLog', async () => {
   console.log(res);
 });
 
-async function slsFetch({
-  protocol = 'https',
-  endpoint,
-  method,
-  path,
-  query,
-  headers: headersAction,
-}: {
-  /**
-   * 协议类型
-   * @default 'https'
-   */
-  protocol?: 'http' | 'https';
-  endpoint: ALIYUN_SLS_ENDPOINT;
-  method?: 'POST' | 'GET';
-  path: string;
-  query?: Record<string, string>;
-  /**
-   * 自定义的 headers
-   */
-  headers?: Record<string, string>;
-}) {
-  const url = new URL(`${protocol}://${endpoint}${path}`);
-  if (query) {
-    for (const [key, value] of Object.entries(query)) {
-      url.searchParams.set(key, value);
-    }
-  }
-  const headersInit = new Headers();
-  if (headersAction) {
-    for (const [key, value] of Object.entries(headersAction)) {
-      headersInit.set(key, value);
-    }
-  }
-  const res = await fetch(url, { method, headers: headersInit });
-  return res.json();
-}
-
-/**
- * doc: https://help.aliyun.com/zh/sls/developer-reference/api-sls-2020-12-30-getlogs
- */
-async function getLogs(
-  { logstore, project }: {
-    logstore: string;
-    /**
-     * Project 名称
-     */
-    project: string;
-    from: UNIX_TIMESTAMP;
-    to: UNIX_TIMESTAMP;
-  },
-  {
-    accessKeyId,
-    secretAccessKey,
-    endpoint,
-  }: {
-    accessKeyId: string;
-    secretAccessKey: string;
-    endpoint: ALIYUN_SLS_ENDPOINT;
-  },
-) {
-  const res = await slsFetch({
-    endpoint,
-    method: 'GET',
-    path: `/logstores/${logstore}`,
-    query: {
-      type: 'log',
-      project,
-    },
-    headers: {
-      'x-acs-action': 'log:GetLogStoreLogs',
-    },
-  });
-  console.log(res);
-  return res;
-}
-
-const project = '';
-const logstore = '';
+const project = Deno.env.get('TEST_SLS_PROJECT')!;
+const logstore = Deno.env.get('TEST_SLS_LOGSTORE')!;
 
 Deno.test('getLogs', async () => {
   const res = await getLogs({
